Initial commit
I have been walking around with this idea to generate music embeddings
from listening history, similar to word2vec, for more than a year now.
I wanted to first finish my music metadata indexer so I could leverage
it, but it is going to take a while. Let's do a quick experiment first,
and see if this idea works at all, because my curiosity is getting the
better of me.