Add scatter plot of embedding
The principal component analysis does not yet work properly -- at least
it looks like it does not. Also, there is something wrong with learning,
where every epoch there is a spike in loss which later settles down. I
suspect that I am not pairing the right labels with the inputs every
time.